==> 0 - generate_DF.ipynb <==
import pandas as pd
from pathlib import Path
from matplotlib import pyplot as plt
import numpy as np

import pandas as pd
from PIL import Image
import xml.etree.ElementTree as ET

def create_DF(files, images_dir):
    """
    Convierte una lista de PosixPaths en un dataframe
    """
    df = pd.DataFrame([])
    for file in files:
        tree = ET.parse(file)
        root = tree.getroot()

        image_dic = []

        for obj in root.findall('object'):
            label = obj.find('name').text       # pineapple
            image = file.stem + ".png"          # fruit0.png
            bndbox = obj.find('bndbox')         
            xmin = float(bndbox.find('xmin').text)      # 38
            ymin = float(bndbox.find('ymin').text)     # 18
            xmax = float(bndbox.find('xmax').text)     # 271)
            ymax = float(bndbox.find('ymax').text)      # 227)

            image_dic.append({
                "label" : label,
                "left" : xmin, 
                "top" : ymin, 
                "right" : xmax,
                "bottom" : ymax
            })
        
        aux_df = pd.DataFrame(image_dic)
        
        # try:
        image_file = next(images_dir.glob(f"**/{file.stem}*.png"))
        aux_df["height"] = float(root.find('size').find('height').text)
        aux_df["width"] = float(root.find('size').find('width').text)
        aux_df["image"] = image_file
        aux_df["image_name"] = root.find('filename').text
        df = pd.concat([df, aux_df])
        # except:
        #     print(file.stem)
    print(df)
    df["center_x"] = (df["right"] + df["left"])/2
    df["center_y"] = (df["bottom"] + df["top"])/2
    df["delta_x"] = df["right"] - df["left"]
    df["delta_y"] = df["bottom"] - df["top"]

    return df

train_files = list(Path("dataset/annotations/").glob("**/*xml"))
print(train_files)

df_train = create_DF(train_files, Path("dataset/images/"))

df_train.head()

valores_unicos_columna2 = df_train['image_name'].unique()

np.random.seed(10)
np.random.shuffle(valores_unicos_columna2)

df_val = df_train[df_train['image_name'].isin(valores_unicos_columna2[:20])]
df_train = df_train[~df_train['image_name'].isin(valores_unicos_columna2[:20])]

images_names = df_train["image_name"].unique()
val_images_name = df_val["image_name"].unique()

train_files_names = images_names
val_files_names = val_images_name

classes = df_train["label"].unique()
classes

classes_idx = {clase:idx for idx,clase in enumerate(classes)}

cant_classes = len(classes)
cant_classes

import pickle
with open("data.pkl","wb") as f:
    pickle.dump((df_train, df_val, train_files_names, val_files_names, classes_idx, cant_classes, classes),f)
