==> 2 - Anchor Boxes.ipynb <==
# Ahora vamos a hacer K-means para elegir los anchor boxes
¿Tiene sentido esto? en función a lo que vimos del dataset?

%load_ext autoreload
%autoreload 2
import pandas as pd
from pathlib import Path
from detection_helper import create_DF
from matplotlib import pyplot as plt
import numpy as np

import pickle
with open("data.pkl","rb") as f:
    (df_train, df_val, train_files_names, val_files_names, classes_idx, cant_classes, classes) = pickle.load(f)

from sklearn.cluster import KMeans

?

clus = KMeans(n_clusters = 4)

clus.fit(df_train[["delta_x", "delta_y"]].values)

clus.cluster_centers_

indexes = np.logical_and((clus.cluster_centers_ < 300)[:,0], (clus.cluster_centers_ < 300)[:,1])

anchor_boxes = clus.cluster_centers_[indexes,:]

anchor_boxes

with open("anchor_boxes.pkl", "wb") as f:
    pickle.dump(anchor_boxes, f)

len(anchor_boxes)

print(2+2)
