==> 4 - Estructura Generador.ipynb <==
# Generador

%load_ext autoreload
%autoreload 2
import numpy as np
import tensorflow as tf
from tensorflow import keras
import pandas as pd
from pathlib import Path
from detection_helper import create_DF
from matplotlib import pyplot as plt
import albumentations as A
from PIL import Image
import matplotlib.patches as patches
import pickle

train_files = list(Path("Dataset/train/").glob("**/*txt"))
val_files = list(Path("Dataset/validation/").glob("**/*txt"))

df_train = create_DF(train_files,Path("Dataset/train/Apple_Grape_Cantaloupe_Watermelon_Pomegranate_Pineapple_Peach_Mango_Banana_Lemon_Tomato_Strawberry_Pear/"))

df_train.to_hdf("df_train.hdf", key="df")

class DataGenerator(keras.utils.Sequence):
    'Generates data for Keras'
    def __init__(self, df, albumentator, anchor_boxes, IoU_alpha=0.5, batch_size=32, crop_size=[300,300], grid_size=[32,32]):
        'Initialization'
        self.df = df.copy()
        self.albumentator = albumentator
        self.anchor_boxes = anchor_boxes
        self.cant_anchor_boxes = len(self.anchor_boxes)
        self.IoU_alpha = IoU_alpha
        self.batch_size = batch_size
        self.crop_size = [300,300]
        self.grid_size = [32,32]
        self.classes = self.df["label"].unique()
        self.classes.sort()
        self.classes_idx = {clase:idx for idx,clase in enumerate(self.classes)}
        self.cant_classes = len(self.classes)
        self.list_IDs = self.df["image_name"].unique()
        self.on_epoch_end()

    def __len__(self):
        'Denotes the number of batches per epoch'
        return int(np.floor(len(self.list_IDs) / self.batch_size))

    def __getitem__(self, index):
        'Generate one batch of data'
        # Generate indexes of the batch
        indexes = self.indexes[index*self.batch_size:(index+1)*self.batch_size]

        # Find list of IDs
        list_IDs_temp = [self.list_IDs[k] for k in indexes]

        # Generate data
        X, y = self.__data_generation(list_IDs_temp)

        return X, y

    def on_epoch_end(self):
        'Updates indexes after each epoch'
        self.indexes = np.arange(len(self.list_IDs))
        np.random.shuffle(self.indexes)

    def __data_generation(self, list_IDs_temp):
        'Generates data containing batch_size samples' # X : (n_samples, *dim, n_channels)
        print(list_IDs_temp)
        return 1,1
#         # Initialization
#         X = np.empty((self.batch_size, *self.dim, self.n_channels))
#         y = np.empty((self.batch_size), dtype=int)

#         # Generate data
#         for i, ID in enumerate(list_IDs_temp):
#             # Store sample
#             X[i,] = np.load('data/' + ID + '.npy')

#             # Store class
#             y[i] = self.labels[ID]

#         return X, keras.utils.to_categorical(y, num_classes=self.n_classes)

with open("anchor_boxes.pkl", "rb") as f:
    anchor_boxes = pickle.load(f)
cant_anchor_boxes = len(anchor_boxes)

transform = A.Compose([
    A.RandomCrop(width=512, height=512),
    A.HorizontalFlip(p=0.5),
    A.RandomBrightnessContrast(p=0.2),
], bbox_params=A.BboxParams(format='coco', min_area=1024, min_visibility=0.1))

with open("data.pkl","rb") as f:
    (df_train, df_val, train_files_names, val_files_names, classes_idx, cant_classes, classes) = pickle.load(f)

gen = DataGenerator(df_train,transform,anchor_boxes=anchor_boxes)

len(gen)

gen[0]
